fix(snap): Fix missing stage-packages
This patch eliminates the following Snapcraft linter warnings:

```
The 'main' part is missing libraries that are not included in the snap or base. They can be satisfied by adding the following entries to the existing stage-packages for this part:
- libgl1
- libglu1-mesa
- libglx0
This part is missing libraries that cannot be satisfied with any available stage-packages known to snapcraft:
- libglut.so.3
These dependencies can be satisfied via additional parts or content sharing. Consider validating configured filesets if this dependency was built.
```

Fixes #10.

Signed-off-by: 林博仁(Buo-ren Lin) <buo.ren.lin@gmail.com>